Add option to delete pods which are in crash loop during StatefulSet update
Emma found this GitHub issue: https://github.com/kubernetes/kubernetes/issues/67250

Essentially, if a pod isn't running successfully, it's never recreated when the StatefulSet pod template is updated - so, if you make a configuration change to fix a crash loop, it'll never get applied unless you delete the remaining pods. This commit allows you to do that through the UI in that circumstance.