Use rows in Deployment table directly instead of maintaining separate queue table
I did this because I was adding environment variables to the deployment config and realized that I'd have to add an encrypted column to the QueuedJob table for them, which sounds like a lot of work, so I moved queueing logic to the Deployment table instead. To pop from the queue, instead of deleting a row, we update the first created row with the "QUEUED" status and set it to "PENDING". As far as I know, locking should work the same with this method, but I haven't done any thorough testing.