Don't panic when log shipper fails to forward signal to child process
I've seen this happen with failing builds. They exit with a status of 0, but the log shipper tries to forward a signal at the same time, resulting in an error like this:

panic: Error relaying signal to process: os: process already finished

goroutine 9 [running]:
main.main.func1()
	/work/main.go:62 +0xbe
created by main.main in goroutine 1
	/work/main.go:55 +0x1f3

This causes the program to exit with a status of 2 (from the Golang panic) instead of the expected 0. It's probably not a fatal error if the process couldn't receive the signal, so we probably shouldn't panic if it happens.